requirements.txt
azureml-core==1.0.65
azure-mgmt-authorization==0.60.0
scikit-learn==0.21.3